moved renderMushroomCards fn into App to pass down as prop
