+++ src/Components/App/App.js
@@ -4,11 +4,27 @@ import './App.css';
 import HomePage from '../HomePage/HomePage'; 
 import ErrorPage from '../ErrorPage/ErrorPage';
 import Nav from '../Nav/Nav'
+import { Link } from 'react-router-dom';
+import { mockMushroomCards } from '../../mockData.js';
+import MushroomCard from "../MushroomCard/MushroomCard";
 
 function App() {
   const [showError, setShowError] = useState(false);
   const [error, setError] = useState(null);
   const [userImage, setUserImage] = useState(null);
+
+  const renderMushroomCards = () => {
+    return mockMushroomCards.map(mushroom => (
+      <Link to={`details/${mushroom.id}`} key={mushroom.id}> 
+        <MushroomCard
+            image={mushroom.image}
+            latinName={mushroom.latinName}
+            commonNames={mushroom.commonNames}
+            probability={mushroom.probability}
+        />
+      </Link>
+    ));
+  };
 
    const handleError = (errorInfo) => {
     setError(errorInfo); 
@@ -20,7 +36,7 @@ function App() {
       <Nav />
       {showError && <ErrorPage error={error} />} 
       <Routes>
-        <Route path="/" element={<HomePage error={error}  userImage={userImage} onImageUpload={setUserImage} />} /> 
+        <Route path="/" element={<HomePage error={error}  userImage={userImage} onImageUpload={setUserImage} renderMushroomCards={renderMushroomCards}/>} /> 
         <Route path="/error" element={<ErrorPage error={error} />} /> 
       </Routes>
     </>
